[Be-Did] Add : RPC Result의 rentEpoch 타입 수정, DiD API 구현 완료 (#81)

+++ server/did/src/main/java/com/c209/did/global/web3/dto/MemcmpDto.java
@@ -7,7 +7,7 @@ import lombok.Builder;
 @Builder
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public record MemcmpDto(
-        @NotNull Long offset,
+        @NotNull Integer offset,
         @NotNull String bytes,
         String encoding
 ) {
